Replace subtraction of Instants and Durations with saturating subtractions
Duration can not be negative, and an underflow causes a panic.

This should fix #12539 as from what I can tell that bug was caused in
`nu-explore::pager::events` from subtracting durations, but I figured
this might be more widespread, and saturating to zero generally makes
sense

I also added the relevant clippy lint to try to prevent this from
happening in the future. I can't think of a reason we would ever want to
subtract durations without checking first.