Deprecate other working-directory-related APIs
Those APIs are re-implemented using `EngineState::cwd()`. None of their behaviors have changed, except for `StateWorkingSet::get_cwd()`, which now returns an empty string instead of panic on error.
This is not an improvement -- I prefer panic over sliently ignoring the error, but it's probably OK since we're deprecating it anyways.