Update eval_filepath() and eval_directory() to the new API
Apparently, arguments that are typed `path` or `directory` are expanded to absolute path before they're even sent to the command.
The `cd` command then trys to expand it again, which feels like an oversight.
I've left the double expansion untouched for now; this commit updates eval_filepath() and eval_directory() to use un-canonicalized PWD,
so that when when type cd while $env.PWD contains a symlink component, we don't accidentally normalize it before the expansion.