Add `ErrSpan` extension trait for `Result`
This adds an extension trait to `Result` that wraps errors in `Spanned`,
saving the effort of calling `.map_err(|err| err.into_spanned(span))`
every time. This will hopefully make it even more likely that someone
will want to use a spanned `io::Error` and make it easier to remove the
impl for `From<io::Error> for ShellError` because that doesn't have span
information.

Discord:
https://discord.com/channels/601130461678272522/855947301380947968/1232142506662297691