Add `to raw` command for making raw input
This is something I've wanted to do a number of times, because
`ExternalStream` is the stream type capable of containing unbounded text
or binary data, and it's also the only way to print raw binary data to
stdout, or a stream of binary data to a file without using `bytes
collect` (with the memory implications of that).

This makes it possible to just convert a list of strings or binaries
into raw input, which can then be used in the same way as if it came
from an external command.